+++ src/main/java/com/paoloamosso/piggus/service/EmailService.java
@@ -1,0 +1,23 @@
+package com.paoloamosso.piggus.service;
+
+import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.mail.SimpleMailMessage;
+import org.springframework.mail.javamail.JavaMailSender;
+import org.springframework.stereotype.Component;
+
+@Component
+public class EmailService {
+
+    @Autowired
+    public JavaMailSender emailSender;
+
+    // TODO refactor Priority 1: Make this send secure and activate security in email configurations
+    public void sendSimpleMessage(
+            String to, String subject, String text) {
+        SimpleMailMessage message = new SimpleMailMessage();
+        message.setTo(to);
+        message.setSubject(subject);
+        message.setText(text);
+        emailSender.send(message);
+    }
+}