+++ src/main/java/com/paoloamosso/piggus/service/ExpenseService.java
@@ -130,12 +130,16 @@ public class ExpenseService {
     }
 
     // Method do synthesize to 2 decimals a double
-    // TODO improvement Priority 4: understand if this is the best way to manage this
+    // TODO improvement Priority 4: understand if this is the best way to manage this, also find another way to check
     public static double round(double value, int places) {
-        if (places < 0) throw new IllegalArgumentException();
+        if (value>-10000) {
+            if (places < 0) throw new IllegalArgumentException();
 
-        BigDecimal bd = new BigDecimal(value);
-        bd = bd.setScale(places, RoundingMode.HALF_UP);
-        return bd.doubleValue();
+            BigDecimal bd = new BigDecimal(value);
+            bd = bd.setScale(places, RoundingMode.HALF_UP);
+            return bd.doubleValue();
+        } else {
+            return value;
+        }
     }
 }