+++ RELEASE NOTES.md
@@ -1,7 +1,22 @@
 PIGGUS - RELEASE NOTES
-=======
+======
+## V1.2.0 - 2018/10/18 ######
+New:
+  * Lost Credentials functionality (first iteration of email sender)
+  * Encrypted email address
+  * Concurrent Session Control
+  * Piggus version available in footer
+  * Deadlines now can be automatically repeated, updated, managed
+---
+Improvements:
+  * Archived deadlines now are correctly shown inside the specific page
+---
+Fixes:
+  * Hidden from git application properties for security reasons
+  * Expense pie does not broke up on first login due to round numbers
+  
 
-## V1.1.0 - 2018/09/XX ######
+## V1.1.0 - 2018/09/22 ######
 New:
   * Homepage doughtnut pie to show kpis
   * First iteration of Privacy & Cookie Policy